+++ packable/packable/src/packable/mod.rs
@@ -6,6 +6,7 @@
 extern crate alloc;
 
 pub mod bounded;
+pub mod map;
 pub mod option;
 pub mod prefix;
 pub mod set;