+++ packable/packable/src/packable/prefix/mod.rs
@@ -5,6 +5,7 @@
 
 mod boxed;
 mod btreeset;
+mod map;
 mod string;
 mod vec;
 
@@ -12,6 +13,7 @@ use core::{convert::Infallible, fmt};
 
 pub use boxed::BoxedSlicePrefix;
 pub use btreeset::BTreeSetPrefix;
+pub use map::{BTreeMapPrefix, HashMapPrefix};
 pub use string::StringPrefix;
 pub use vec::VecPrefix;